Refactor Game class to use Story class for story management
